% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bv.standardization.R
\name{bv.standardization}
\alias{bv.standardization}
\title{BV standardization}
\usage{
bv.standardization(population, mean.target = 100, var.target = 10,
  gen = NULL, database = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{mean.target}{Target mean}

\item{var.target}{Target variance}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\description{
Function to get mean and genetic variance of a trait to a fixed value
}
