% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeOriginsR.R
\name{codeOriginsR}
\alias{codeOriginsR}
\title{Origins-coding(R)}
\usage{
codeOriginsR(M)
}
\arguments{
\item{M}{Origins matrix}
}
\value{
Bit-wise coded origins
}
\description{
R-Version of the internal bitwise-coding of origins
}
