% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating.phenotypic.transform.R
\name{creating.phenotypic.transform}
\alias{creating.phenotypic.transform}
\title{Calculate breeding values}
\usage{
creating.phenotypic.transform(population,
  phenotypic.transform.function = NULL, trait = 1)
}
\arguments{
\item{population}{Population list}

\item{phenotypic.transform.function}{Phenotypic transformation to apply}

\item{trait}{Trait for which a transformation is to be applied
data(ex_pop)
trafo <- function(x){ return(x^2)}
ex_pop <- creating.phenotypic.transform(ex_pop, phenotypic.transform.function=trafo)}
}
\value{
Population-list with a new phenotypic transformation function
}
\description{
Internal function to calculate the breeding value of a given individual
}
