% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi}
\alias{epi}
\title{Martini-Test function}
\usage{
epi(y, Z, G = NULL)
}
\arguments{
\item{y}{y}

\item{Z}{genomic information matrix}

\item{G}{kinship matrix}
}
\value{
Estimated breeding values
}
\description{
Internal function to perform martini test
}
