% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cullingtime.R
\name{get.cullingtime}
\alias{get.cullingtime}
\title{Derive time of culling}
\usage{
get.cullingtime(population, database = NULL, gen = NULL,
  cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\description{
Function to devide the time of culling for all individuals
}
