% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.death.point.R
\name{get.death.point}
\alias{get.death.point}
\title{Derive death point}
\usage{
get.death.point(population, database = NULL, gen = NULL,
  cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\value{
Time of death of in gen/database/cohorts selected individuals
}
\description{
Function to devide the time of death for each individual (NA for individuals that are still alive))
}
\examples{
data(ex_pop)
get.death.point(ex_pop, gen=2)
}
