% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.time.point.R
\name{get.time.point}
\alias{get.time.point}
\title{Derive time point}
\usage{
get.time.point(population, database = NULL, gen = NULL,
  cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\description{
Function to devide time point for each individual
}
