% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.bve.R
\name{insert.bve}
\alias{insert.bve}
\title{Export estimated breeding values}
\usage{
insert.bve(population, bves, type = "bve", count = 1)
}
\arguments{
\item{population}{Population list}

\item{bves}{Matrix of breeding values to enter (one row per individual with 1 element coding individual name)}

\item{type}{which time of values to input (default: "bve", alt: "bv", "pheno")}

\item{count}{Counting for economic cost calculation (default: 1 - (one observation (for "pheno"), one genotyping (for "bve")))}
}
\value{
Population-List with newly entered estimated breeding values
}
\description{
Function to export estimated breeding values
}
\examples{
data(ex_pop)
bv <- get.bv(ex_pop, gen=2)
new.bve <- cbind( colnames(bv), bv[,1]) ## Unrealistic but you do not get better than this!
ex_pop <- insert.bve(ex_pop, bves=new.bve)
}
