% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.emp.store.R
\name{kinship.exp.store}
\alias{kinship.exp.store}
\title{Derive expected kinship}
\usage{
kinship.exp.store(population, gen = NULL, database = NULL,
  cohorts = NULL, depth.pedigree = 7, start.kinship = NULL,
  elements = NULL, mult = NULL, storage_save = 3)
}
\arguments{
\item{population}{Population list}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{depth.pedigree}{Depth of the pedigree in generations}

\item{start.kinship}{Relationship matrix of the individuals in the first considered generation}
}
\description{
Function to derive expected kinship
}
