% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.population.R
\name{plot.population}
\alias{plot.population}
\title{Plot Population}
\usage{
\method{plot}{population}(object, type = "bve", gen = NULL,
  database = NULL, cohorts = NULL, ...)
}
\arguments{
\item{object}{Population-list}

\item{...}{additional arguments affecting the summary produced}
}
\value{
Summary of the population list including number of individuals, genone length and trait overview
}
\description{
Basic plot of the population list
}
\examples{
data(ex_pop)
plot(ex_pop)
}
