% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.dataset.R
\name{sheep_chip}
\alias{sheep_chip}
\title{sheep chip}
\source{
Prieur et al
}
\usage{
sheep_chip
}
\description{
Genome for sheep according to Prieur et al.
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
