% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortd.R
\name{sortd}
\alias{sortd}
\title{Apply sort and unique}
\usage{
sortd(v)
}
\arguments{
\item{v}{Vector}
}
\value{
numerical sorted vector without duplicates
}
\description{
Efficient function to perform sort(unique(v))
}
