% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.database.R
\name{get.database}
\alias{get.database}
\title{gen/database/cohorts conversion}
\usage{
get.database(population, gen = NULL, database = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\description{
Function to derive a database based on gen/database/cohorts
}
